refactor: Period Enum 위치를 global.type 패키지로 이동
