+++ deokhugam/src/main/java/com/deokhugam/deokhugam_server/domain/book/dto/response/PopularBookDto.java
@@ -1,6 +1,6 @@
 package com.deokhugam.deokhugam_server.domain.book.dto.response;
 
-import com.deokhugam.deokhugam_server.domain.user.dto.response.Period;
+import com.deokhugam.deokhugam_server.global.type.Period;
 import java.time.LocalDateTime;
 import java.util.UUID;
