+++ deokhugam/src/main/java/com/deokhugam/deokhugam_server/domain/user/controller/UserController.java
@@ -3,9 +3,8 @@ package com.deokhugam.deokhugam_server.domain.user.controller;
 import com.deokhugam.deokhugam_server.domain.user.dto.request.UserLoginRequest;
 import com.deokhugam.deokhugam_server.domain.user.dto.request.UserRegisterRequest;
 import com.deokhugam.deokhugam_server.domain.user.dto.request.UserUpdateRequest;
-import com.deokhugam.deokhugam_server.domain.user.dto.response.Period;
+import com.deokhugam.deokhugam_server.global.type.Period;
 import com.deokhugam.deokhugam_server.domain.user.dto.response.UserDto;
-import com.deokhugam.deokhugam_server.domain.user.entity.User;
 import com.deokhugam.deokhugam_server.domain.user.service.UserService;
 import jakarta.validation.Valid;
 import java.util.List;
