--- deokhugam/src/main/java/com/deokhugam/deokhugam_server/domain/user/dto/response/Period.java
@@ -1,3 +1,0 @@
-package com.deokhugam.deokhugam_server.domain.user.dto.response;
-
-public enum Period {DAILY, WEEKLY, MONTHLY, ALL_TIME}
