+++ deokhugam/src/main/java/com/deokhugam/deokhugam_server/domain/user/dto/response/PowerUserDto.java
@@ -1,5 +1,6 @@
 package com.deokhugam.deokhugam_server.domain.user.dto.response;
 
+import com.deokhugam.deokhugam_server.global.type.Period;
 import java.time.LocalDateTime;
 import java.util.UUID;
